requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pdfminer.six>=20231228
PyPDF2>=3.0.0
ruyaml>=0.91.0
tqdm>=4.66.4
